Removed dependency on controlfx.
The problem was that the breadcrumb bar did not shrink if space was short.
I created a new breadcrumb bar that lead to an error in javaFX (I created a bugreport)
The only dependency was SegmentedButton but that was so easy to copy. In fact
it is only a HBox with togglebuttons and the correct css is already in the jdk.